requirements.txt
netaddr==0.7.12
python-nmap==0.3.4
netifaces==0.10.4
psutil==2.1.1
